requirements.txt
requests>=2.12
GitPython>=2.1.11
simplejson>=3.16.0